fix: swagger advice 종속성 문제 해결 (버전 업그레이드)
