+++ src/main/kotlin/kr/hhplus/be/server/global/exception/ErrorResponse.kt
@@ -1,0 +1,11 @@
+package kr.hhplus.be.server.global.exception
+
+import java.time.LocalDateTime
+
+data class ErrorResponse(
+    val timestamp: LocalDateTime,
+    val status: Int,
+    val error: String,
+    val message: String,
+    val path: String
+)
